fix accounts crash
fixes #4541
introduced in
https://github.com/PrismLauncher/PrismLauncher/commit/3c46d8a412956a759f61ae802c540ef72d00b35d
The original commit introduced m_name but never used it.
When the endActivity would be called with a count of 0 this would crash
the laucnher.
How to reproduce: try to switch your skin in quick succession until you
get 429 too many requests as response to the login.
I will also presume this is not the only crash caused by this(hopefully
it is as it was not catched for four years)

Signed-off-by: Trial97 <alexandru.tripon97@gmail.com>
(cherry picked from commit f4b22dae90363f71a3203439111ad1596215b751)